Completed Motion Blur Image
Moving to a different format for the project. Instead of having
a generic `Demo` trait implemented by all the demos amongst which
one can switch by pressing 1-0 keys, We'll just have one demo only.
And, To switch between different images, They can checkout to
respective branch

+++ src/types/ray.rs
@@ -3,11 +3,11 @@ use crate::types::Vec3;
 pub struct Ray {
     a: Vec3,
     b: Vec3,
-    time: f32,
+    time: f64,
 }
 
 impl Ray {
-    pub fn new(a: Vec3, b: Vec3, time: f32) -> Ray {
+    pub fn new(a: Vec3, b: Vec3, time: f64) -> Ray {
         Ray { a, b, time }
     }
     #[inline]
@@ -19,11 +19,11 @@ impl Ray {
         self.b
     }
     #[inline]
-    pub fn point_at_parameter(&self, t: f32) -> Vec3 {
+    pub fn point_at_parameter(&self, t: f64) -> Vec3 {
         self.a + self.b * t
     }
     #[inline]
-    pub const fn time(&self) -> f32 {
+    pub const fn time(&self) -> f64 {
         self.time
     }
 }
